+++ src/app/bases/EntityBase.ts
@@ -25,8 +25,16 @@ export abstract class EntityBase extends Model {
 
     /**
      * [ObjectionJS] Array of primary column names.
+     * Should be overriden (['id', 'tenant_id']) for composite PK.
      */
     public static readonly idColumn: string[] = ['id']
+
+    /**
+     * Same with `idColumn`, but transform snakeCase to camelCase.
+     */
+    public static get idProp(): string[] {
+        return this.idColumn.map<string>(camelCase)
+    }
 
     /**
      * An array of non-primary unique column names.
@@ -34,11 +42,10 @@ export abstract class EntityBase extends Model {
     public static readonly uniqColumn: string[] = []
 
     /**
-     * Same with `idColumn`, but transform snakeCase to camelCase.
-     * Should be overriden (['id', 'tenantId']) for composite PK.
+     * Same with `uniqColumn`, but transform snakeCase to camelCase.
      */
-    public static get idProp(): string[] {
-        return this.idColumn.map<string>(camelCase)
+    public static get uniqProp(): string[] {
+        return this.uniqColumn.map<string>(camelCase)
     }